Improve Open In apps settings
