+++ src/renderer/src/components/settings/GeneralPane.tsx
@@ -1,23 +1,14 @@
-/* eslint-disable max-lines -- Why: GeneralPane is the single owner of all general settings UI;
-   splitting individual settings into separate files would scatter related controls without a
-   meaningful abstraction boundary. */
-import { useEffect, useRef, useState } from 'react'
-import type { GlobalSettings, OpenInApplication } from '../../../../shared/types'
-import { Button } from '../ui/button'
-import { Input } from '../ui/input'
-import { Label } from '../ui/label'
-import { Separator } from '../ui/separator'
-import { Download, FolderOpen, Loader2, RefreshCw, Star, Timer } from 'lucide-react'
+import type React from 'react'
+import type { GlobalSettings } from '../../../../shared/types'
 import { useAppStore } from '../../store'
+import { Separator } from '../ui/separator'
 import { CliSection } from './CliSection'
-import { toast } from 'sonner'
-import {
-  DEFAULT_EDITOR_AUTO_SAVE_DELAY_MS,
-  MAX_EDITOR_AUTO_SAVE_DELAY_MS,
-  MIN_EDITOR_AUTO_SAVE_DELAY_MS
-} from '../../../../shared/constants'
-import { OPEN_IN_APPLICATIONS_MAX } from '../../../../shared/open-in-applications'
-import { clampNumber } from '@/lib/terminal-theme'
+import { GeneralCacheTimerSection } from './GeneralCacheTimerSection'
+import { GeneralEditorSettingsSection } from './GeneralEditorSettingsSection'
+import { GeneralNetworkSettingsSection } from './GeneralNetworkSettingsSection'
+import { GeneralSupportSection } from './GeneralSupportSection'
+import { GeneralUpdateSettingsSection } from './GeneralUpdateSettingsSection'
+import { GeneralWorkspaceSettingsSection } from './GeneralWorkspaceSettingsSection'
 import {
   GENERAL_CACHE_TIMER_SEARCH_ENTRIES,
   GENERAL_CLI_SEARCH_ENTRIES,
@@ -29,42 +20,25 @@ import {
   GENERAL_UPDATE_SEARCH_ENTRIES,
   GENERAL_WORKSPACE_SEARCH_ENTRIES
 } from './general-search'
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../ui/select'
 import { RecentTabOrderControl } from './RecentTabOrderControl'
-import { SearchableSetting } from './SearchableSetting'
 import { matchesSettingsSearch } from './settings-search'
-import {
-  SettingsSegmentedControl,
-  SettingsSubsectionHeader,
-  SettingsSwitch,
-  SettingsSwitchRow
-} from './SettingsFormControls'
-import { useMountedRef } from '@/hooks/useMountedRef'
-import { normalizeProxyBypassRules, normalizeProxyUrl } from '../../../../shared/network-proxy'
+import { SettingsSubsectionHeader } from './SettingsFormControls'
 
-function createOpenInApplication(): OpenInApplication {
-  return {
-    id:
-      globalThis.crypto?.randomUUID?.() ??
-      `open-in-${Date.now().toString(36)}-${Math.random().toString(36).slice(2)}`,
-    label: '',
-    command: ''
-  }
-}
-
-function createPresetOpenInApplication(label: string, command: string): OpenInApplication {
-  return {
-    ...createOpenInApplication(),
-    label,
-    command
-  }
-}
-
-export function shouldCommitOpenInApplicationsDraft(applications: OpenInApplication[]): boolean {
-  return applications.every((application) => {
-    return application.label.trim() !== '' && application.command.trim() !== ''
-  })
-}
+export {
+  createAutoSaveDelayDraftState,
+  updateAutoSaveDelayDraftState,
+  type AutoSaveDelayDraftState
+} from './GeneralEditorSettingsSection'
+export {
+  createHttpProxyBypassRulesDraftState,
+  createHttpProxyUrlDraftState,
+  setHttpProxyUrlDraftErrorState,
+  updateHttpProxyBypassRulesDraftState,
+  updateHttpProxyUrlDraftState,
+  type HttpProxyBypassRulesDraftState,
+  type HttpProxyUrlDraftState
+} from './GeneralNetworkSettingsSection'
+export { shouldCommitOpenInApplicationsDraft } from './OpenInMenuSetting'
 
 export function getDesktopPlatformFromUserAgent(userAgent: string): 'darwin' | 'win32' | 'other' {
   if (userAgent.includes('Mac')) {
@@ -77,151 +51,6 @@ export function getDesktopPlatformFromUserAgent(userAgent: string): 'darwin' | '
 }
 
 export { GENERAL_PANE_SEARCH_ENTRIES }
-
-export type AutoSaveDelayDraftState = {
-  sourceDelayMs: number
-  draft: string
-}
-
-export function createAutoSaveDelayDraftState(
-  editorAutoSaveDelayMs: number
-): AutoSaveDelayDraftState {
-  return {
-    sourceDelayMs: editorAutoSaveDelayMs,
-    draft: String(editorAutoSaveDelayMs)
-  }
-}
-
-function resolveAutoSaveDelayDraftState(
-  state: AutoSaveDelayDraftState,
-  editorAutoSaveDelayMs: number
-): AutoSaveDelayDraftState {
-  return state.sourceDelayMs === editorAutoSaveDelayMs
-    ? state
-    : createAutoSaveDelayDraftState(editorAutoSaveDelayMs)
-}
-
-export function updateAutoSaveDelayDraftState(
-  state: AutoSaveDelayDraftState,
-  editorAutoSaveDelayMs: number,
-  draft: string
-): AutoSaveDelayDraftState {
-  return {
-    // Why: settings persistence is async, so a committed draft must stay tied
-    // to the current source until the persisted value reloads.
-    ...resolveAutoSaveDelayDraftState(state, editorAutoSaveDelayMs),
-    draft
-  }
-}
-
-export type HttpProxyUrlDraftState = {
-  sourceValue: string
-  draft: string
-  error: string | null
-}
-
-export function createHttpProxyUrlDraftState(
-  httpProxyUrl: string | undefined
-): HttpProxyUrlDraftState {
-  const sourceValue = httpProxyUrl ?? ''
-  return {
-    sourceValue,
-    draft: sourceValue,
-    error: null
-  }
-}
-
-function resolveHttpProxyUrlDraftState(
-  state: HttpProxyUrlDraftState,
-  httpProxyUrl: string | undefined
-): HttpProxyUrlDraftState {
-  const sourceValue = httpProxyUrl ?? ''
-  return state.sourceValue === sourceValue ? state : createHttpProxyUrlDraftState(httpProxyUrl)
-}
-
-export function updateHttpProxyUrlDraftState(
-  state: HttpProxyUrlDraftState,
-  httpProxyUrl: string | undefined,
-  draft: string
-): HttpProxyUrlDraftState {
-  return {
-    // Why: settings persistence is async, so edits after an external settings
-    // reload must build on the latest persisted proxy source.
-    ...resolveHttpProxyUrlDraftState(state, httpProxyUrl),
-    draft,
-    error: null
-  }
-}
-
-export function setHttpProxyUrlDraftErrorState(
-  state: HttpProxyUrlDraftState,
-  httpProxyUrl: string | undefined,
-  error: string
-): HttpProxyUrlDraftState {
-  return {
-    ...resolveHttpProxyUrlDraftState(state, httpProxyUrl),
-    error
-  }
-}
-
-export type HttpProxyBypassRulesDraftState = {
-  sourceValue: string
-  draft: string
-}
-
-export function createHttpProxyBypassRulesDraftState(
-  httpProxyBypassRules: string | undefined
-): HttpProxyBypassRulesDraftState {
-  const sourceValue = httpProxyBypassRules ?? ''
-  return {
-    sourceValue,
-    draft: sourceValue
-  }
-}
-
-function resolveHttpProxyBypassRulesDraftState(
-  state: HttpProxyBypassRulesDraftState,
-  httpProxyBypassRules: string | undefined
-): HttpProxyBypassRulesDraftState {
-  const sourceValue = httpProxyBypassRules ?? ''
-  return state.sourceValue === sourceValue
-    ? state
-    : createHttpProxyBypassRulesDraftState(httpProxyBypassRules)
-}
-
-export function updateHttpProxyBypassRulesDraftState(
-  state: HttpProxyBypassRulesDraftState,
-  httpProxyBypassRules: string | undefined,
-  draft: string
-): HttpProxyBypassRulesDraftState {
-  return {
-    ...resolveHttpProxyBypassRulesDraftState(state, httpProxyBypassRules),
-    draft
-  }
-}
-
-type OpenInApplicationsDraftState = {
-  sourceApplications: OpenInApplication[] | undefined
-  draft: OpenInApplication[]
-}
-
-function createOpenInApplicationsDraftState(
-  openInApplications: OpenInApplication[] | undefined
-): OpenInApplicationsDraftState {
-  return {
-    sourceApplications: openInApplications,
-    draft: openInApplications ?? []
-  }
-}
-
-function resolveOpenInApplicationsDraftState(
-  state: OpenInApplicationsDraftState,
-  openInApplications: OpenInApplication[] | undefined
-): OpenInApplicationsDraftState {
-  return state.sourceApplications === openInApplications
-    ? state
-    : createOpenInApplicationsDraftState(openInApplications)
-}
 
 type GeneralPaneProps = {
   settings: GlobalSettings
@@ -239,249 +68,6 @@ export function GeneralPane({
   wslCapabilitiesLoading
 }: GeneralPaneProps): React.JSX.Element {
   const searchQuery = useAppStore((s) => s.settingsSearchQuery)
-  const updateStatus = useAppStore((s) => s.updateStatus)
-  const mountedRef = useMountedRef()
-  // Why: the 'error' variant of UpdateStatus does not carry a `version` field.
-  // The main process emits `{ state: 'error' }` for both check failures (no
-  // version known yet) and download/install failures (version was known from
-  // the preceding 'available'/'downloading'/'downloaded' state). Cache the
-  // last-known version so the error copy below can distinguish the two cases
-  // without adding IPC. Mirrors `versionRef` in UpdateCard.tsx.
-  const updateVersionRef = useRef<string | null>(null)
-  if (
-    (updateStatus.state === 'available' ||
-      updateStatus.state === 'downloading' ||
-      updateStatus.state === 'downloaded') &&
-    updateStatus.version
-  ) {
-    updateVersionRef.current = updateStatus.version
-  } else if (
-    updateStatus.state === 'checking' ||
-    updateStatus.state === 'idle' ||
-    updateStatus.state === 'not-available'
-  ) {
-    // Why: a new check cycle has started or completed cleanly. Clear the
-    // cached version so a subsequent check failure cannot be mis-classified
-    // as a download failure based on a stale version from a prior cycle.
-    updateVersionRef.current = null
-  }
-  const [appVersion, setAppVersion] = useState<string | null>(null)
-  const [autoSaveDelayDraftState, setAutoSaveDelayDraftState] = useState(() =>
-    createAutoSaveDelayDraftState(settings.editorAutoSaveDelayMs)
-  )
-  const [httpProxyUrlDraftState, setHttpProxyUrlDraftState] = useState(() =>
-    createHttpProxyUrlDraftState(settings.httpProxyUrl)
-  )
-  const [httpProxyBypassRulesDraftState, setHttpProxyBypassRulesDraftState] = useState(() =>
-    createHttpProxyBypassRulesDraftState(settings.httpProxyBypassRules)
-  )
-  const [openInApplicationsDraftState, setOpenInApplicationsDraftState] = useState(() =>
-    createOpenInApplicationsDraftState(settings.openInApplications)
-  )
-  // Why: the star state is derived from gh, not from settings, so it does not
-  // live in the global settings store. 'hidden' covers the gh-unavailable and
-  // already-starred-on-a-previous-session cases so the section drops out for
-  // users who can't or don't need to act.
-  //
-  // We start in 'loading' and render a placeholder at the exact same
-  // dimensions as the resolved section. When gh resolves to 'hidden', the
-  // placeholder collapses with a grid-rows transition so content above it
-  // doesn't shift; anything below (nothing today, but future-proof) eases up.
-  const [starState, setStarState] = useState<
-    'loading' | 'not-starred' | 'starred' | 'starring' | 'hidden' | 'error'
-  >('loading')
-
-  useEffect(() => {
-    let cancelled = false
-    void window.api.updater.getVersion().then((version) => {
-      if (!cancelled) {
-        setAppVersion(version)
-      }
-    })
-    return () => {
-      cancelled = true
-    }
-  }, [])
-
-  useEffect(() => {
-    let cancelled = false
-    void window.api.gh.checkOrcaStarred().then((result) => {
-      if (cancelled) {
-        return
-      }
-      if (result === null) {
-        setStarState('hidden')
-      } else {
-        setStarState(result ? 'starred' : 'not-starred')
-      }
-    })
-    return () => {
-      cancelled = true
-    }
-  }, [])
-
-  const handleStarClick = async (): Promise<void> => {
-    if (starState !== 'not-starred' && starState !== 'error') {
-      return
-    }
-    setStarState('starring')
-    const ok = await window.api.gh.starOrca('settings')
-    if (!ok) {
-      if (mountedRef.current) {
-        setStarState('error')
-      }
-      return
-    }
-    if (mountedRef.current) {
-      setStarState('starred')
-    }
-    // Why: clicking star anywhere should also permanently mute the
-    // threshold-based nag so the user isn't re-prompted via the popup.
-    await window.api.starNag.complete()
-  }
-
-  const resolvedAutoSaveDelayDraftState = resolveAutoSaveDelayDraftState(
-    autoSaveDelayDraftState,
-    settings.editorAutoSaveDelayMs
-  )
-  if (resolvedAutoSaveDelayDraftState !== autoSaveDelayDraftState) {
-    // Why: Settings can be updated outside this pane; reconcile drafts before
-    // paint so the visible input never lags behind the persisted value.
-    setAutoSaveDelayDraftState(resolvedAutoSaveDelayDraftState)
-  }
-  const autoSaveDelayDraft = resolvedAutoSaveDelayDraftState.draft
-  const updateAutoSaveDelayDraft = (draft: string): void => {
-    setAutoSaveDelayDraftState((current) =>
-      updateAutoSaveDelayDraftState(current, settings.editorAutoSaveDelayMs, draft)
-    )
-  }
-
-  const resolvedOpenInApplicationsDraftState = resolveOpenInApplicationsDraftState(
-    openInApplicationsDraftState,
-    settings.openInApplications
-  )
-  if (resolvedOpenInApplicationsDraftState !== openInApplicationsDraftState) {
-    // Why: the Open In rows are a local draft, but Settings can reload them
-    // externally; sync before paint instead of after an Effect pass.
-    setOpenInApplicationsDraftState(resolvedOpenInApplicationsDraftState)
-  }
-  const openInApplicationsDraft = resolvedOpenInApplicationsDraftState.draft
-  const updateOpenInApplicationsDraft = (draft: OpenInApplication[]): void => {
-    setOpenInApplicationsDraftState((current) => ({
-      ...resolveOpenInApplicationsDraftState(current, settings.openInApplications),
-      draft
-    }))
-  }
-
-  const resolvedHttpProxyUrlDraftState = resolveHttpProxyUrlDraftState(
-    httpProxyUrlDraftState,
-    settings.httpProxyUrl
-  )
-  if (resolvedHttpProxyUrlDraftState !== httpProxyUrlDraftState) {
-    // Why: Settings can change outside this pane; reconcile the proxy draft
-    // before paint so stale network values do not briefly appear.
-    setHttpProxyUrlDraftState(resolvedHttpProxyUrlDraftState)
-  }
-  const httpProxyUrlDraft = resolvedHttpProxyUrlDraftState.draft
-  const httpProxyUrlError = resolvedHttpProxyUrlDraftState.error
-  const updateHttpProxyUrlDraft = (draft: string): void => {
-    setHttpProxyUrlDraftState((current) =>
-      updateHttpProxyUrlDraftState(current, settings.httpProxyUrl, draft)
-    )
-  }
-
-  const resolvedHttpProxyBypassRulesDraftState = resolveHttpProxyBypassRulesDraftState(
-    httpProxyBypassRulesDraftState,
-    settings.httpProxyBypassRules
-  )
-  if (resolvedHttpProxyBypassRulesDraftState !== httpProxyBypassRulesDraftState) {
-    // Why: Proxy bypass rules are local input state, but settings reloads can
-    // replace their source while this pane is mounted.
-    setHttpProxyBypassRulesDraftState(resolvedHttpProxyBypassRulesDraftState)
-  }
-  const httpProxyBypassRulesDraft = resolvedHttpProxyBypassRulesDraftState.draft
-  const updateHttpProxyBypassRulesDraft = (draft: string): void => {
-    setHttpProxyBypassRulesDraftState((current) =>
-      updateHttpProxyBypassRulesDraftState(current, settings.httpProxyBypassRules, draft)
-    )
-  }
-
-  const commitOpenInApplications = (applications: OpenInApplication[]): void => {
-    if (!shouldCommitOpenInApplicationsDraft(applications)) {
-      return
-    }
-    updateSettings({ openInApplications: applications })
-  }
-
-  const applyOpenInApplicationsDraft = (applications: OpenInApplication[]): void => {
-    updateOpenInApplicationsDraft(applications)
-    commitOpenInApplications(applications)
-  }
-
-  const handleBrowseWorkspace = async () => {
-    const path = await window.api.repos.pickFolder()
-    if (path) {
-      updateSettings({ workspaceDir: path })
-    }
-  }
-
-  const commitAutoSaveDelay = (): void => {
-    const trimmed = autoSaveDelayDraft.trim()
-    if (trimmed === '') {
-      setAutoSaveDelayDraftState(createAutoSaveDelayDraftState(settings.editorAutoSaveDelayMs))
-      return
-    }
-
-    const value = Number(trimmed)
-    if (!Number.isFinite(value)) {
-      setAutoSaveDelayDraftState(createAutoSaveDelayDraftState(settings.editorAutoSaveDelayMs))
-      return
-    }
-
-    const next = clampNumber(
-      Math.round(value),
-      MIN_EDITOR_AUTO_SAVE_DELAY_MS,
-      MAX_EDITOR_AUTO_SAVE_DELAY_MS
-    )
-    updateSettings({ editorAutoSaveDelayMs: next })
-    setAutoSaveDelayDraftState((current) =>
-      updateAutoSaveDelayDraftState(current, settings.editorAutoSaveDelayMs, String(next))
-    )
-  }
-
-  const commitHttpProxyUrl = (): void => {
-    const normalized = normalizeProxyUrl(httpProxyUrlDraft)
-    if (!normalized.ok) {
-      setHttpProxyUrlDraftState((current) =>
-        setHttpProxyUrlDraftErrorState(current, settings.httpProxyUrl, normalized.message)
-      )
-      return
-    }
-    setHttpProxyUrlDraftState((current) =>
-      updateHttpProxyUrlDraftState(current, settings.httpProxyUrl, normalized.value)
-    )
-    if (normalized.value !== (settings.httpProxyUrl ?? '')) {
-      updateSettings({ httpProxyUrl: normalized.value })
-    }
-  }
-
-  const commitHttpProxyBypassRules = (): void => {
-    const normalized = normalizeProxyBypassRules(httpProxyBypassRulesDraft)
-    setHttpProxyBypassRulesDraftState((current) =>
-      updateHttpProxyBypassRulesDraftState(current, settings.httpProxyBypassRules, normalized)
-    )
-    if (normalized !== (settings.httpProxyBypassRules ?? '')) {
-      updateSettings({ httpProxyBypassRules: normalized })
-    }
-  }
-
-  const handleRestartToUpdate = (): void => {
-    // Why: quitAndInstall resolves immediately (the actual quit happens in a
-    // deferred timer in the main process), so rejection here is only possible
-    // if the IPC channel itself breaks. Log defensively; the user will notice
-    // the app didn't restart and can retry.
-    void window.api.updater.quitAndInstall().catch(console.error)
-  }
 
   const visibleSections = [
     matchesSettingsSearch(searchQuery, GENERAL_NAVIGATION_SEARCH_ENTRIES) ? (
@@ -499,408 +85,25 @@ export function GeneralPane({
       </section>
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_WORKSPACE_SEARCH_ENTRIES) ? (
-      <section key="workspace" className="space-y-4">
-        <SettingsSubsectionHeader
-          title="Workspace"
-          description="Configure where new workspaces are created."
-        />
-
-        <SearchableSetting
-          title="Workspace Directory"
-          description="Root directory where workspace folders are created."
-          keywords={['workspace', 'folder', 'path', 'worktree']}
-          className="space-y-2"
-        >
-          <Label>Workspace Directory</Label>
-          <div className="flex gap-2">
-            <Input
-              value={settings.workspaceDir}
-              onChange={(e) => updateSettings({ workspaceDir: e.target.value })}
-              className="flex-1 text-xs"
-            />
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleBrowseWorkspace}
-              className="shrink-0 gap-1.5"
-            >
-              <FolderOpen className="size-3.5" />
-              Browse
-            </Button>
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Root directory where workspace folders are created.
-          </p>
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Nest Workspaces"
-          description="Create workspaces inside a repo-named subfolder."
-          keywords={['nested', 'subfolder', 'directory']}
-        >
-          <SettingsSwitchRow
-            label="Nest Workspaces"
-            description="Create workspaces inside a repo-named subfolder."
-            checked={settings.nestWorkspaces}
-            onChange={() => updateSettings({ nestWorkspaces: !settings.nestWorkspaces })}
-          />
-        </SearchableSetting>
-
-        {/* Why: the "Don't ask again" toast in the delete-worktree dialog
-            deep-links here, so the wrapper id must stay stable. Renaming it
-            breaks that toast action even though this pane still renders fine. */}
-        <div id="general-skip-delete-worktree-confirm" className="scroll-mt-6">
-          <SearchableSetting
-            title="Ask Before Deleting Workspaces"
-            description="Show a confirmation dialog before deleting a workspace."
-            keywords={['delete', 'worktree', 'confirm', 'dialog', 'skip', 'prompt']}
-          >
-            <SettingsSwitchRow
-              label="Ask Before Deleting Workspaces"
-              description="Show a confirmation before deleting a workspace from the context menu. Failed deletes still surface a Force Delete fallback."
-              checked={!settings.skipDeleteWorktreeConfirm}
-              onChange={() =>
-                updateSettings({
-                  skipDeleteWorktreeConfirm: !settings.skipDeleteWorktreeConfirm
-                })
-              }
-            />
-          </SearchableSetting>
-        </div>
-
-        <div id="general-skip-delete-automation-confirm" className="scroll-mt-6">
-          <SearchableSetting
-            title="Ask Before Deleting Automations"
-            description="Show a confirmation dialog before deleting an automation and its run history."
-            keywords={['delete', 'automation', 'confirm', 'dialog', 'skip', 'prompt']}
-          >
-            <SettingsSwitchRow
-              label="Ask Before Deleting Automations"
-              description="Show a confirmation before deleting automations and their run history."
-              checked={!settings.skipDeleteAutomationConfirm}
-              onChange={() =>
-                updateSettings({
-                  skipDeleteAutomationConfirm: !settings.skipDeleteAutomationConfirm
-                })
-              }
-            />
-          </SearchableSetting>
-        </div>
-
-        <SearchableSetting
-          title="Open In Menu"
-          description="Add custom launchers to the workspace Open in menu."
-          keywords={['open in', 'editor', 'launcher', 'cursor', 'zed', 'command', 'vscode']}
-          className="space-y-3"
-        >
-          <div className="space-y-1">
-            <Label>Open In Menu</Label>
-            <p className="text-xs text-muted-foreground">
-              VS Code is always included first. Add executables to show extra entries in each
-              workspace&apos;s Open in menu.
-            </p>
-            <p className="text-xs text-muted-foreground">
-              Commands are not shell-parsed. Use only an executable command name. For flags, use a
-              wrapper script.
-            </p>
-          </div>
-          <div className="flex flex-wrap gap-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() =>
-                applyOpenInApplicationsDraft([
-                  ...openInApplicationsDraft,
-                  createPresetOpenInApplication('Cursor', 'cursor')
-                ])
-              }
-            >
-              Add Cursor
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() =>
-                applyOpenInApplicationsDraft([
-                  ...openInApplicationsDraft,
-                  createPresetOpenInApplication('Zed', 'zed')
-                ])
-              }
-            >
-              Add Zed
-            </Button>
-          </div>
-          <div className="space-y-2">
-            {openInApplicationsDraft.map((app, index) => (
-              <div key={app.id} className="grid grid-cols-1 gap-2 sm:grid-cols-[1fr_1fr_auto]">
-                <Input
-                  value={app.label}
-                  placeholder="Label"
-                  onChange={(event) => {
-                    const next = [...openInApplicationsDraft]
-                    next[index] = { ...app, label: event.target.value }
-                    updateOpenInApplicationsDraft(next)
-                  }}
-                  onBlur={() => commitOpenInApplications(openInApplicationsDraft)}
-                  onKeyDown={(event) => {
-                    if (event.key === 'Enter') {
-                      commitOpenInApplications(openInApplicationsDraft)
-                    }
-                  }}
-                />
-                <Input
-                  value={app.command}
-                  placeholder="Executable command"
-                  onChange={(event) => {
-                    const next = [...openInApplicationsDraft]
-                    next[index] = { ...app, command: event.target.value }
-                    updateOpenInApplicationsDraft(next)
-                  }}
-                  onBlur={() => commitOpenInApplications(openInApplicationsDraft)}
-                  onKeyDown={(event) => {
-                    if (event.key === 'Enter') {
-                      commitOpenInApplications(openInApplicationsDraft)
-                    }
-                  }}
-                />
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={() => {
-                    const next = openInApplicationsDraft.filter((entry) => entry.id !== app.id)
-                    updateOpenInApplicationsDraft(next)
-                    commitOpenInApplications(next)
-                  }}
-                >
-                  Remove
-                </Button>
-              </div>
-            ))}
-          </div>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() =>
-              updateOpenInApplicationsDraft([...openInApplicationsDraft, createOpenInApplication()])
-            }
-            disabled={openInApplicationsDraft.length >= OPEN_IN_APPLICATIONS_MAX}
-          >
-            Add Custom Launcher
-          </Button>
-        </SearchableSetting>
-      </section>
+      <GeneralWorkspaceSettingsSection
+        key="workspace"
+        settings={settings}
+        updateSettings={updateSettings}
+      />
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_NETWORK_SEARCH_ENTRIES) ? (
-      <section key="network" className="space-y-4">
-        <SettingsSubsectionHeader
-          title="Network"
-          description="Configure app-level network routing."
-        />
-
-        <SearchableSetting
-          title="HTTP Proxy"
-          description="Proxy URL for Orca network requests and local terminal children."
-          keywords={['proxy', 'http_proxy', 'https_proxy', 'network', 'dock', 'launchpad']}
-          className="space-y-3"
-        >
-          <div className="space-y-1">
-            <Label htmlFor="settings-http-proxy-url">HTTP Proxy</Label>
-            <p className="text-xs text-muted-foreground">
-              Leave empty to use system proxy settings and inherited proxy environment variables.
-            </p>
-          </div>
-          <Input
-            id="settings-http-proxy-url"
-            value={httpProxyUrlDraft}
-            onChange={(e) => {
-              updateHttpProxyUrlDraft(e.target.value)
-            }}
-            onBlur={commitHttpProxyUrl}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') {
-                e.currentTarget.blur()
-              }
-            }}
-            placeholder="http://proxy.example.com:8080"
-            autoCapitalize="none"
-            autoCorrect="off"
-            autoComplete="off"
-            spellCheck={false}
-            aria-invalid={httpProxyUrlError ? true : undefined}
-            className="font-mono text-xs"
-          />
-          {httpProxyUrlError ? (
-            <p className="text-xs text-destructive">{httpProxyUrlError}</p>
-          ) : (
-            <p className="text-xs text-muted-foreground">
-              Supports http, https, socks, socks4, and socks5 URLs.
-            </p>
-          )}
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Proxy Bypass Rules"
-          description="Hosts that should bypass the configured HTTP proxy."
-          keywords={['proxy', 'bypass', 'no_proxy', 'localhost', 'network']}
-          className="space-y-3"
-        >
-          <div className="space-y-1">
-            <Label htmlFor="settings-http-proxy-bypass-rules">Proxy Bypass Rules</Label>
-            <p className="text-xs text-muted-foreground">
-              Optional. Separate hosts with commas, semicolons, or new lines.
-            </p>
-          </div>
-          <Input
-            id="settings-http-proxy-bypass-rules"
-            value={httpProxyBypassRulesDraft}
-            onChange={(e) => updateHttpProxyBypassRulesDraft(e.target.value)}
-            onBlur={commitHttpProxyBypassRules}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') {
-                e.currentTarget.blur()
-              }
-            }}
-            placeholder="localhost, 127.0.0.1, *.internal"
-            autoCapitalize="none"
-            autoCorrect="off"
-            autoComplete="off"
-            spellCheck={false}
-            className="font-mono text-xs"
-          />
-        </SearchableSetting>
-      </section>
+      <GeneralNetworkSettingsSection
+        key="network"
+        settings={settings}
+        updateSettings={updateSettings}
+      />
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_EDITOR_SEARCH_ENTRIES) ? (
-      <section key="editor" className="space-y-4">
-        <SettingsSubsectionHeader
-          title="Editor"
-          description="Configure how Orca persists file edits."
-        />
-
-        <SearchableSetting
-          title="Auto Save Files"
-          description="Save editor and editable diff changes automatically after a short pause."
-          keywords={['autosave', 'save']}
-        >
-          <SettingsSwitchRow
-            label="Auto Save Files"
-            description="Save editor and editable diff changes automatically after a short pause."
-            checked={settings.editorAutoSave}
-            onChange={() => updateSettings({ editorAutoSave: !settings.editorAutoSave })}
-          />
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Auto Save Delay"
-          description="How long Orca waits after your last edit before saving automatically."
-          keywords={['autosave', 'delay', 'milliseconds']}
-          className="flex items-center justify-between gap-4 py-2"
-        >
-          <div className="min-w-0 flex-1 space-y-0.5">
-            <Label>Auto Save Delay</Label>
-            <p className="text-xs text-muted-foreground">
-              How long Orca waits after your last edit before saving automatically. First launch
-              defaults to {DEFAULT_EDITOR_AUTO_SAVE_DELAY_MS} ms.
-            </p>
-          </div>
-          <div className="flex shrink-0 items-center gap-2">
-            <Input
-              type="number"
-              min={MIN_EDITOR_AUTO_SAVE_DELAY_MS}
-              max={MAX_EDITOR_AUTO_SAVE_DELAY_MS}
-              step={250}
-              value={autoSaveDelayDraft}
-              onChange={(e) => updateAutoSaveDelayDraft(e.target.value)}
-              onBlur={commitAutoSaveDelay}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  commitAutoSaveDelay()
-                }
-              }}
-              className="number-input-clean w-28 text-right tabular-nums"
-            />
-            <span className="text-xs text-muted-foreground">ms</span>
-          </div>
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Default Diff View"
-          description="Preferred presentation format for showing git diffs by default."
-          keywords={['diff', 'view', 'inline', 'side-by-side', 'split']}
-          className="flex items-center justify-between gap-4 py-2"
-        >
-          <div className="min-w-0 flex-1 space-y-0.5">
-            <Label>Default Diff View</Label>
-            <p className="text-xs text-muted-foreground">
-              Preferred presentation format for showing git diffs by default.
-            </p>
-          </div>
-          <SettingsSegmentedControl
-            ariaLabel="Default Diff View"
-            value={settings.diffDefaultView}
-            onChange={(option) => updateSettings({ diffDefaultView: option })}
-            options={[
-              { value: 'inline', label: 'Inline' },
-              { value: 'side-by-side', label: 'Side-by-side' }
-            ]}
-          />
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Default Diff File Tree"
-          description="Show or hide the file tree when opening combined diff views."
-          keywords={['diff', 'tree', 'file tree', 'combined diff', 'sidebar']}
-          className="flex items-center justify-between gap-4 py-2"
-        >
-          <div className="min-w-0 flex-1 space-y-0.5">
-            <Label>Default Diff File Tree</Label>
-            <p className="text-xs text-muted-foreground">
-              Show or hide the file tree when opening combined diff views.
-            </p>
-          </div>
-          <SettingsSegmentedControl
-            ariaLabel="Default Diff File Tree"
-            value={settings.combinedDiffFileTreeVisibleByDefault ? 'shown' : 'hidden'}
-            onChange={(option) =>
-              updateSettings({ combinedDiffFileTreeVisibleByDefault: option === 'shown' })
-            }
-            options={[
-              { value: 'shown', label: 'Shown' },
-              { value: 'hidden', label: 'Hidden' }
-            ]}
-          />
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Minimap"
-          description="Show the minimap overview when editing a file."
-          keywords={['minimap', 'overview', 'code', 'scroll']}
-        >
-          <SettingsSwitchRow
-            label="Minimap"
-            description="Show the minimap overview when editing a file."
-            checked={settings.editorMinimapEnabled}
-            onChange={() =>
-              updateSettings({ editorMinimapEnabled: !settings.editorMinimapEnabled })
-            }
-          />
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="Markdown Review Notes"
-          description="Show local markdown review note controls in rich editor mode."
-          keywords={['markdown', 'review', 'notes', 'annotations', 'agents']}
-        >
-          <SettingsSwitchRow
-            label="Markdown Review Notes"
-            description="Show local markdown note controls in rich editor mode and agent handoff actions."
-            checked={settings.markdownReviewToolsEnabled}
-            onChange={() =>
-              updateSettings({ markdownReviewToolsEnabled: !settings.markdownReviewToolsEnabled })
-            }
-          />
-        </SearchableSetting>
-      </section>
+      <GeneralEditorSettingsSection
+        key="editor"
+        settings={settings}
+        updateSettings={updateSettings}
+      />
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_CLI_SEARCH_ENTRIES) ? (
       <CliSection
@@ -914,184 +117,14 @@ export function GeneralPane({
       />
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_CACHE_TIMER_SEARCH_ENTRIES) ? (
-      <section key="cache-timer" className="space-y-4">
-        <SettingsSubsectionHeader
-          title="Prompt Cache Timer"
-          description="Claude caches your conversation to reduce costs. When idle too long the cache expires and the next message resends full context at higher cost. This shows a countdown so you know when to resume."
-        />
-
-        <SearchableSetting
-          title="Cache Timer"
-          description="Show a countdown after a Claude agent becomes idle."
-          keywords={GENERAL_CACHE_TIMER_SEARCH_ENTRIES.flatMap((entry) => [
-            entry.title,
-            entry.description ?? '',
-            ...(entry.keywords ?? [])
-          ])}
-          className="flex items-center justify-between gap-4 py-2"
-        >
-          <div className="min-w-0 flex-1 space-y-0.5">
-            <div className="flex items-center gap-2">
-              <Timer className="size-4 text-muted-foreground" />
-              <Label>Cache Timer</Label>
-            </div>
-            <p className="text-xs text-muted-foreground">
-              Show a countdown in the sidebar after a Claude agent becomes idle.
-            </p>
-          </div>
-          <SettingsSwitch
-            ariaLabel="Cache Timer"
-            checked={settings.promptCacheTimerEnabled}
-            onChange={() => {
-              const enabling = !settings.promptCacheTimerEnabled
-              updateSettings({ promptCacheTimerEnabled: enabling })
-              if (enabling) {
-                useAppStore.getState().seedCacheTimersForIdleTabs()
-              }
-            }}
-          />
-        </SearchableSetting>
-
-        {settings.promptCacheTimerEnabled && (
-          <SearchableSetting
-            title="Timer Duration"
-            description="Match this to your provider's cache TTL."
-            keywords={['cache', 'timer', 'duration', 'ttl']}
-            className="flex items-center justify-between gap-4 py-2 pl-7"
-          >
-            <div className="min-w-0 flex-1 space-y-0.5">
-              <Label>Timer Duration</Label>
-              <p className="text-xs text-muted-foreground">
-                Match this to your provider&apos;s cache TTL. The default is 5 minutes.
-              </p>
-            </div>
-            <Select
-              value={String(settings.promptCacheTtlMs)}
-              onValueChange={(v) => updateSettings({ promptCacheTtlMs: Number(v) })}
-            >
-              <SelectTrigger size="sm" className="h-7 text-xs w-[120px]">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="300000">5 minutes</SelectItem>
-                <SelectItem value="3600000">1 hour</SelectItem>
-              </SelectContent>
-            </Select>
-          </SearchableSetting>
-        )}
-      </section>
+      <GeneralCacheTimerSection
+        key="cache-timer"
+        settings={settings}
+        updateSettings={updateSettings}
+      />
     ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_UPDATE_SEARCH_ENTRIES) ? (
-      <section key="updates" className="space-y-4">
-        <SettingsSubsectionHeader
-          title="Updates"
-          description={`Current version: ${appVersion ?? '…'}`}
-        />
-
-        <SearchableSetting
-          title="Check for Updates"
-          description="Check for app updates and install a newer Orca version."
-          keywords={['update', 'version', 'release notes', 'download']}
-          className="space-y-3"
-        >
-          <div className="flex items-center gap-3">
-            <Button
-              variant="outline"
-              size="sm"
-              // Why: Shift-click opts this check into the release-candidate
-              // channel. Keep the affordance hidden — it's a power-user
-              // shortcut, not a discoverable toggle.
-              onClick={(event) =>
-                window.api.updater.check({
-                  includePrerelease: event.shiftKey
-                })
-              }
-              disabled={updateStatus.state === 'checking' || updateStatus.state === 'downloading'}
-              className="gap-2"
-            >
-              {updateStatus.state === 'checking' ? (
-                <Loader2 className="size-3.5 animate-spin" />
-              ) : (
-                <RefreshCw className="size-3.5" />
-              )}
-              Check for Updates
-            </Button>
-
-            {updateStatus.state === 'available' ? (
-              <Button
-                variant="default"
-                size="sm"
-                onClick={() => {
-                  void window.api.updater.download().catch((error) => {
-                    toast.error('Could not start the update download.', {
-                      description: String((error as Error)?.message ?? error)
-                    })
-                  })
-                }}
-                className="gap-2"
-              >
-                <Download className="size-3.5" />
-                Install Update ({updateStatus.version})
-              </Button>
-            ) : updateStatus.state === 'downloaded' ? (
-              <Button variant="default" size="sm" onClick={handleRestartToUpdate} className="gap-2">
-                <Download className="size-3.5" />
-                Restart to Update ({updateStatus.version})
-              </Button>
-            ) : null}
-          </div>
-
-          <p className="text-xs text-muted-foreground">
-            {updateStatus.state === 'idle' && 'Updates are checked automatically on launch.'}
-            {updateStatus.state === 'checking' && 'Checking for updates...'}
-            {updateStatus.state === 'available' && (
-              <>
-                Version {updateStatus.version} is available. Click &quot;Install Update&quot; to
-                download and install it.{' '}
-                <a
-                  href={
-                    updateStatus.releaseUrl ??
-                    `https://github.com/stablyai/orca/releases/tag/v${updateStatus.version}`
-                  }
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="underline hover:text-foreground"
-                >
-                  Release notes
-                </a>
-              </>
-            )}
-            {updateStatus.state === 'not-available' && 'You\u2019re on the latest version.'}
-            {updateStatus.state === 'downloading' &&
-              `Downloading v${updateStatus.version}... ${updateStatus.percent}%`}
-            {updateStatus.state === 'downloaded' && (
-              <>
-                Version {updateStatus.version} is ready to install.{' '}
-                <a
-                  href={
-                    updateStatus.releaseUrl ??
-                    `https://github.com/stablyai/orca/releases/tag/v${updateStatus.version}`
-                  }
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="underline hover:text-foreground"
-                >
-                  Release notes
-                </a>
-              </>
-            )}
-            {updateStatus.state === 'error' &&
-              // Why: `{ state: 'error' }` is emitted for both check-time
-              // failures (no version cached) and download/install failures
-              // (version cached from a prior 'available'/'downloading'/
-              // 'downloaded' state). Label accordingly so a download failure
-              // isn't mislabeled as a "check" failure. Mirrors UpdateCard.tsx.
-              (updateVersionRef.current
-                ? `Update error. ${updateStatus.message}`
-                : `Update check failed. ${updateStatus.message}`)}
-          </p>
-        </SearchableSetting>
-      </section>
+      <GeneralUpdateSettingsSection key="updates" />
     ) : null
     // Note: the Support section is rendered outside this array so it can own
     // its own loading placeholder and its own collapsing Separator. Without
@@ -1108,124 +141,8 @@ export function GeneralPane({
         </div>
       ))}
       {matchesSettingsSearch(searchQuery, GENERAL_SUPPORT_SEARCH_ENTRIES) ? (
-        <SupportSection
-          state={starState}
-          hasPrecedingSections={visibleSections.length > 0}
-          onStarClick={handleStarClick}
-        />
+        <GeneralSupportSection hasPrecedingSections={visibleSections.length > 0} />
       ) : null}
-    </div>
-  )
-}
-
-type SupportSectionProps = {
-  state: 'loading' | 'not-starred' | 'starring' | 'starred' | 'hidden' | 'error'
-  hasPrecedingSections: boolean
-  onStarClick: () => void | Promise<void>
-}
-
-function SupportSection({
-  state,
-  hasPrecedingSections,
-  onStarClick
-}: SupportSectionProps): React.JSX.Element {
-  // Why: 'hidden' means gh is unavailable or the user had already starred on a
-  // previous session — in both cases we collapse the entire section (including
-  // its leading Separator) so the settings pane doesn't carry an empty strip.
-  // For every other state we render the full row so the initial layout is
-  // stable: the skeleton-to-live swap happens in place and a post-click
-  // "Starred" confirmation does not shift anything above or below it.
-  const collapsed = state === 'hidden'
-
-  return (
-    <section
-      className={`grid transition-[grid-template-rows,opacity] duration-300 ease-out ${
-        collapsed ? 'grid-rows-[0fr] opacity-0' : 'grid-rows-[1fr] opacity-100'
-      }`}
-      aria-hidden={collapsed}
-    >
-      <div className="min-h-0 overflow-hidden">
-        <div className="space-y-8">
-          {hasPrecedingSections ? <Separator /> : null}
-          <div className="space-y-4">
-            <SettingsSubsectionHeader title="Support Orca" />
-            {state === 'loading' ? <SupportRowSkeleton /> : null}
-            {state !== 'loading' && state !== 'hidden' ? (
-              <SupportRow state={state} onStarClick={onStarClick} />
-            ) : null}
-          </div>
-        </div>
-      </div>
-    </section>
-  )
-}
-
-function SupportRowSkeleton(): React.JSX.Element {
-  return (
-    <div className="flex items-center justify-between gap-4 py-2" aria-hidden="true">
-      <div className="h-4 w-36 rounded bg-muted/50 animate-pulse" />
-      <div className="h-8 w-24 rounded-md bg-muted/50 animate-pulse" />
-    </div>
-  )
-}
-
-function SupportRow({
-  state,
-  onStarClick
-}: {
-  state: 'not-starred' | 'starring' | 'starred' | 'error'
-  onStarClick: () => void | Promise<void>
-}): React.JSX.Element {
-  // Why: the left-hand label is the setting's identity and must not change
-  // when the user clicks — the row should still read "Star Orca on GitHub"
-  // afterwards. The right-hand control is what changes: before starring it
-  // is a button; after a successful star we swap in a small inline "Thanks"
-  // confirmation so the row keeps the same shape without showing a stale,
-  // disabled button.
-  return (
-    <SearchableSetting
-      title="Star Orca on GitHub"
-      description="Support the project with a GitHub star via the gh CLI."
-      keywords={['star', 'github', 'support', 'feedback', 'like']}
-      className="flex items-center justify-between gap-4 py-2"
-    >
-      <Label>Star Orca on GitHub</Label>
-      {state === 'starred' ? (
-        <SupportRowThanks />
-      ) : (
-        <Button
-          variant="default"
-          size="sm"
-          onClick={() => void onStarClick()}
-          disabled={state === 'starring'}
-          className="shrink-0 gap-1.5"
-        >
-          {state === 'starring' ? (
-            <Loader2 className="size-3.5 animate-spin" />
-          ) : (
-            <Star className="size-3.5" />
-          )}
-          {state === 'starring' ? 'Starring…' : state === 'error' ? 'Try Again' : 'Star'}
-        </Button>
-      )}
-    </SearchableSetting>
-  )
-}
-
-function SupportRowThanks(): React.JSX.Element {
-  // Why: match the size="sm" button's h-8 / gap-1.5 / px-3 dimensions so the
-  // row height stays identical when the button is swapped out. Without the
-  // fixed height, the text baseline collapses ~6px and the entire row
-  // shrinks, shifting everything below.
-  return (
-    <div
-      className="shrink-0 inline-flex h-8 items-center gap-1.5 px-3 text-sm font-medium
-        text-amber-400/90 animate-in fade-in slide-in-from-right-1 duration-300"
-      role="status"
-      aria-live="polite"
-    >
-      <Star className="size-3.5 fill-amber-400/80 text-amber-400/80" aria-hidden="true" />
-      Thanks for the support!
     </div>
   )
 }
