+++ src/renderer/src/components/settings/GeneralSupportSection.tsx
@@ -1,0 +1,180 @@
+import type React from 'react'
+import { useEffect, useState } from 'react'
+import { Loader2, Star } from 'lucide-react'
+import { useMountedRef } from '@/hooks/useMountedRef'
+import { Button } from '../ui/button'
+import { Label } from '../ui/label'
+import { Separator } from '../ui/separator'
+import { SearchableSetting } from './SearchableSetting'
+import { SettingsSubsectionHeader } from './SettingsFormControls'
+
+type SupportState = 'loading' | 'not-starred' | 'starring' | 'starred' | 'hidden' | 'error'
+
+type GeneralSupportSectionProps = {
+  hasPrecedingSections: boolean
+}
+
+export function GeneralSupportSection({
+  hasPrecedingSections
+}: GeneralSupportSectionProps): React.JSX.Element {
+  const mountedRef = useMountedRef()
+  // Why: the star state is derived from gh, not from settings, so it does not
+  // live in the global settings store. 'hidden' covers the gh-unavailable and
+  // already-starred-on-a-previous-session cases so the section drops out for
+  // users who can't or don't need to act.
+  //
+  // We start in 'loading' and render a placeholder at the exact same
+  // dimensions as the resolved section. When gh resolves to 'hidden', the
+  // placeholder collapses with a grid-rows transition so content above it
+  // doesn't shift; anything below (nothing today, but future-proof) eases up.
+  const [starState, setStarState] = useState<SupportState>('loading')
+
+  useEffect(() => {
+    let cancelled = false
+    void window.api.gh.checkOrcaStarred().then((result) => {
+      if (cancelled) {
+        return
+      }
+      if (result === null) {
+        setStarState('hidden')
+      } else {
+        setStarState(result ? 'starred' : 'not-starred')
+      }
+    })
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  const handleStarClick = async (): Promise<void> => {
+    if (starState !== 'not-starred' && starState !== 'error') {
+      return
+    }
+    setStarState('starring')
+    const ok = await window.api.gh.starOrca('settings')
+    if (!ok) {
+      if (mountedRef.current) {
+        setStarState('error')
+      }
+      return
+    }
+    if (mountedRef.current) {
+      setStarState('starred')
+    }
+    // Why: clicking star anywhere should also permanently mute the
+    // threshold-based nag so the user isn't re-prompted via the popup.
+    await window.api.starNag.complete()
+  }
+
+  return (
+    <SupportSection
+      state={starState}
+      hasPrecedingSections={hasPrecedingSections}
+      onStarClick={handleStarClick}
+    />
+  )
+}
+
+type SupportSectionProps = {
+  state: SupportState
+  hasPrecedingSections: boolean
+  onStarClick: () => void | Promise<void>
+}
+
+function SupportSection({
+  state,
+  hasPrecedingSections,
+  onStarClick
+}: SupportSectionProps): React.JSX.Element {
+  // Why: 'hidden' means gh is unavailable or the user had already starred on a
+  // previous session. Collapse the whole section, including its leading
+  // Separator, so the settings pane doesn't carry an empty strip.
+  const collapsed = state === 'hidden'
+
+  return (
+    <section
+      className={`grid transition-[grid-template-rows,opacity] duration-300 ease-out ${
+        collapsed ? 'grid-rows-[0fr] opacity-0' : 'grid-rows-[1fr] opacity-100'
+      }`}
+      aria-hidden={collapsed}
+    >
+      <div className="min-h-0 overflow-hidden">
+        <div className="space-y-8">
+          {hasPrecedingSections ? <Separator /> : null}
+          <div className="space-y-4">
+            <SettingsSubsectionHeader title="Support Orca" />
+            {state === 'loading' ? <SupportRowSkeleton /> : null}
+            {state !== 'loading' && state !== 'hidden' ? (
+              <SupportRow state={state} onStarClick={onStarClick} />
+            ) : null}
+          </div>
+        </div>
+      </div>
+    </section>
+  )
+}
+
+function SupportRowSkeleton(): React.JSX.Element {
+  return (
+    <div className="flex items-center justify-between gap-4 py-2" aria-hidden="true">
+      <div className="h-4 w-36 rounded bg-muted/50 animate-pulse" />
+      <div className="h-8 w-24 rounded-md bg-muted/50 animate-pulse" />
+    </div>
+  )
+}
+
+function SupportRow({
+  state,
+  onStarClick
+}: {
+  state: 'not-starred' | 'starring' | 'starred' | 'error'
+  onStarClick: () => void | Promise<void>
+}): React.JSX.Element {
+  // Why: the left-hand label is the setting's identity and must not change
+  // when the user clicks. The right-hand control is what changes: before
+  // starring it is a button; after success it becomes a small confirmation.
+  return (
+    <SearchableSetting
+      title="Star Orca on GitHub"
+      description="Support the project with a GitHub star via the gh CLI."
+      keywords={['star', 'github', 'support', 'feedback', 'like']}
+      className="flex items-center justify-between gap-4 py-2"
+    >
+      <Label>Star Orca on GitHub</Label>
+      {state === 'starred' ? (
+        <SupportRowThanks />
+      ) : (
+        <Button
+          variant="default"
+          size="sm"
+          onClick={() => void onStarClick()}
+          disabled={state === 'starring'}
+          className="shrink-0 gap-1.5"
+        >
+          {state === 'starring' ? (
+            <Loader2 className="size-3.5 animate-spin" />
+          ) : (
+            <Star className="size-3.5" />
+          )}
+          {state === 'starring' ? 'Starring...' : state === 'error' ? 'Try Again' : 'Star'}
+        </Button>
+      )}
+    </SearchableSetting>
+  )
+}
+
+function SupportRowThanks(): React.JSX.Element {
+  // Why: match the size="sm" button's h-8 / gap-1.5 / px-3 dimensions so the
+  // row height stays identical when the button is swapped out.
+  return (
+    <div
+      className="shrink-0 inline-flex h-8 items-center gap-1.5 px-3 text-sm font-medium
+        text-amber-400/90 animate-in fade-in slide-in-from-right-1 duration-300"
+      role="status"
+      aria-live="polite"
+    >
+      <Star className="size-3.5 fill-amber-400/80 text-amber-400/80" aria-hidden="true" />
+      Thanks for the support!
+    </div>
+  )
+}
