+++ src/renderer/src/components/settings/OpenInMenuSetting.test.ts
@@ -1,0 +1,63 @@
+import { describe, expect, it } from 'vitest'
+import {
+  createPresetOpenInApplication,
+  shouldCommitOpenInApplicationsDraft
+} from './OpenInMenuSetting'
+import { isOpenInAppPresetAdded, OPEN_IN_APP_PRESETS } from '@/lib/open-in-app-catalog'
+import type { OpenInAppPreset } from '@/lib/open-in-app-catalog'
+
+function requirePreset(id: string): OpenInAppPreset {
+  const preset = OPEN_IN_APP_PRESETS.find((entry) => entry.id === id)
+  if (!preset) {
+    throw new Error(`Preset not found: ${id}`)
+  }
+  return preset
+}
+
+describe('OpenInMenuSetting presets', () => {
+  it('creates stable preset rows for known apps', () => {
+    const cursor = requirePreset('cursor')
+
+    expect(createPresetOpenInApplication(cursor)).toEqual({
+      id: 'cursor',
+      label: 'Cursor',
+      command: 'cursor'
+    })
+  })
+
+  it('recognizes legacy preset rows by command', () => {
+    const cursor = requirePreset('cursor')
+
+    expect(isOpenInAppPresetAdded([{ command: ' cursor ' }], cursor)).toBe(true)
+  })
+})
+
+describe('OpenInMenuSetting application drafts', () => {
+  it('does not commit rows until both label and command are present', () => {
+    expect(
+      shouldCommitOpenInApplicationsDraft([{ id: 'draft', label: 'Cursor', command: '' }])
+    ).toBe(false)
+    expect(
+      shouldCommitOpenInApplicationsDraft([{ id: 'draft', label: '', command: 'cursor' }])
+    ).toBe(false)
+    expect(
+      shouldCommitOpenInApplicationsDraft([{ id: 'draft', label: '   ', command: 'cursor' }])
+    ).toBe(false)
+    expect(
+      shouldCommitOpenInApplicationsDraft([{ id: 'draft', label: 'Cursor', command: '   ' }])
+    ).toBe(false)
+  })
+
+  it('allows commit when every draft row has a label and command', () => {
+    expect(shouldCommitOpenInApplicationsDraft([])).toBe(true)
+    expect(
+      shouldCommitOpenInApplicationsDraft([{ id: 'cursor', label: 'Cursor', command: 'cursor' }])
+    ).toBe(true)
+    expect(
+      shouldCommitOpenInApplicationsDraft([
+        { id: 'cursor', label: 'Cursor', command: 'cursor' },
+        { id: 'zed', label: 'Zed', command: 'zed' }
+      ])
+    ).toBe(true)
+  })
+})
